Переделан экран акций
