+++ Finance/Presentation/Home/Presenter/HomePresenter.swift
@@ -17,7 +17,7 @@ final class HomePresenter {
     private let databaseService = DatabaseService.shared
     private let networkService = NetworkService.shared
     private let service = [ChoiceService(name: "Курс валют", img: "exchange-rate.png", vc: ExchangeRateViewController()),
-                           ChoiceService(name: "Акции", img: "stock-market.png", vc: StocksController())]
+                           ChoiceService(name: "Акции", img: "stock-market.png", vc: StocksViewController())]
 
     private var lastTransaction = [LastTransaction]()
 //    private var lastTransaction = [LastTransaction]() {
