+++ Finance/Presentation/Stocks/View/StocksOutput.swift
@@ -1,0 +1,12 @@
+//
+//  StocksOutput.swift
+//  Finance
+//
+//  Created by Александр Меренков on 20.06.2023.
+//
+
+import Foundation
+
+protocol StocksOutput: AnyObject {
+    func viewIsReady()
+}
